Unify test suites
