+++ jit/tests/collapse-adjacent-rules.test.js
@@ -4,7 +4,7 @@ const fs = require('fs')
 const path = require('path')
 
 function run(input, config = {}) {
-  return postcss([tailwind(config)]).process(input, { from: path.resolve(__filename) })
+  return postcss(tailwind(config)).process(input, { from: path.resolve(__filename) })
 }
 
 test('collapse adjacent rules', () => {
@@ -50,6 +50,6 @@ test('collapse adjacent rules', () => {
     let expectedPath = path.resolve(__dirname, './collapse-adjacent-rules.test.css')
     let expected = fs.readFileSync(expectedPath, 'utf8')
 
-    expect(result.css).toMatchCss(expected)
+    expect(result.css).toMatchFormattedCss(expected)
   })
 })
